move `replaceObject` to core
