+++ packages/@tailwindcss-upgrade/src/codemods/template/migrate-modernize-arbitrary-values.ts
@@ -4,8 +4,8 @@ import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
 import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
 import { computeVariantSignature } from '../../../../tailwindcss/src/signatures'
 import { isPositiveInteger } from '../../../../tailwindcss/src/utils/infer-data-type'
+import { replaceObject } from '../../../../tailwindcss/src/utils/replace-object'
 import * as ValueParser from '../../../../tailwindcss/src/value-parser'
-import { replaceObject } from '../../utils/replace-object'
 import { walkVariants } from '../../utils/walk-variants'
 
 export function migrateModernizeArbitraryValues(
