+++ packages/tailwindcss/src/utils/replace-object.ts
